requirements.txt
flask==3.1.1
waitress==3.0.2
anthropic==0.52.0
google-api-python-client==2.166.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
gspread==6.2.1
requests==2.32.3
python-dotenv==1.1.0
apscheduler==3.11.0
openai==1.82.0
reportlab==4.4.10
pytest==8.3.4
